yandex_parser/requirements.txt
pandas==1.3.5
Flask~=2.0.2
Werkzeug~=2.0.2
requests==2.27.1
python-dotenv~=0.19.1
gevent==21.12.0
Flask-RESTful==0.3.9
gunicorn==20.1.0
python-anticaptcha==1.0.0
selenium==4.1.3
beautifulsoup4==4.11.1
lxml==4.8.0
pymongo
2captcha-python
dnspython